+++ packages/mel-components/src/utils/useOptionsApi.ts
@@ -3,8 +3,7 @@ import { useRequestReactive, wheneverEffectScopeImmediate } from '@mixte/use';
 import { isFunction, isPlainObject } from 'mixte';
 import { computed, toValue, watch } from 'vue';
 
-// #region OptionsApi
-
+// #region OptionsApiRequest
 export type OptionsApiRequest<T> = (...args: any[]) => Promise<
   T[] |
   {
@@ -14,7 +13,8 @@ export type OptionsApiRequest<T> = (...args: any[]) => Promise<
       };
   }
 >;
-
+// #endregion OptionsApiRequest
+// #region OptionsApiConfig
 export interface OptionsApiConfig<T> {
   /** 请求选项数据源的方法 */
   api?: OptionsApiRequest<T>;
@@ -31,9 +31,9 @@ export interface OptionsApiConfig<T> {
    */
   immediate?: boolean;
 }
-
+// #endregion OptionsApiConfig
+// #region OptionsApi
 export type OptionsApi<T> = OptionsApiRequest<T> | OptionsApiConfig<T>;
-
 // #endregion OptionsApi
 
 /**
@@ -51,15 +51,25 @@ function extractData<T>(res: any): T[] | undefined {
   }
 }
 
-export function useOptionsApi<T>(optionsApi: Ref<OptionsApi<T> | undefined>) {
+interface UseOptionsApiOptions {
+  /** 请求参数 */
+  params?: MaybeRefOrGetter<Record<string, any> | undefined>;
+}
+
+export function useOptionsApi<T>(
+  optionsApi: Ref<OptionsApi<T> | undefined>,
+  options?: UseOptionsApiOptions,
+) {
   const propApi = computed(() => isFunction(optionsApi.value) ? optionsApi.value : optionsApi.value?.api);
   const propImmediate = computed(() => isFunction(optionsApi.value) ? true : (optionsApi.value?.immediate ?? true));
   const propParams = computed(() => isFunction(optionsApi.value) ? undefined : toValue(optionsApi.value?.params));
 
+  const optionParams = computed(() => toValue(options?.params));
+
+  const params = computed(() => Array.isArray(propParams.value) ? propParams.value : { ...propParams.value, ...optionParams.value });
+
   const api = useRequestReactive<T[] | undefined>(async () => {
-    const params = propParams.value == null ? [] : [propParams.value];
-    const response = await propApi.value?.(...params);
-    return extractData<T>(response);
+    return extractData<T>(await propApi.value?.(params.value));
   });
 
   const loading = computed(() => propApi.value && api.isLoading);
@@ -69,7 +79,7 @@ export function useOptionsApi<T>(optionsApi: Ref<OptionsApi<T> | undefined>) {
   }
 
   wheneverEffectScopeImmediate(propApi, () => {
-    watch(() => JSON.stringify(propParams.value), () => {
+    watch(() => JSON.stringify(params.value), () => {
       api.execute();
     });
   });