+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
   - 🌟 [@mixte/mel-components] 组件 `MelSelect` 新增 `filterOptionMethod` 参数对数据源选项进行筛选
+  - 🌟 [@mixte/mel-components] 组件 `MelSelect` 新增 `optionsApi.remoteKey` 参数以支持筛选时使用远程搜索
   - 🐞 [@mixte/components] 修复 `GridTable` 在现代渲染模式使用 `bordered` 时, 表头分割线还存在的问题 *
 
 ## [v3.5.0-beta.14]